backend codes updated icon updated

+++ src/app/utls/actions/delete/deleteProjects.ts
@@ -10,6 +10,10 @@ const deleteProject = async (id: string) => {
     
     try {
         const session = await getServerSession(authOptions);
+
+        if (session?.user?.role !== "Admin") {
+            throw new Error("Forbidden: Only admins can delete");
+        }
         
         const res = await fetch(`${process.env.BASE_URL}/projects/${id}`, {
         method: "DELETE",
